+++ crates/ruma-client-api/src/r0/message.rs
@@ -1,4 +1,4 @@
-//! Enpoints for sending and receiving messages
+//! Endpoints for sending and receiving messages
 
 pub mod get_message_events;
 pub mod send_message_event;