+++ crates/ruma-serde/src/strings.rs
@@ -32,7 +32,7 @@ where
 }
 
 /// Serde serializiation decorator to map None to an empty String,
-/// and forward Somes to the Serialize implemention for T.
+/// and forward Somes to the Serialize implementation for T.
 ///
 /// To be used like this:
 /// `#[serde(serialize_with = "empty_string_as_none")]`